+++ src/settings.py
@@ -44,6 +44,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'rest_framework',
     'corsheaders',
+    'django_nose',
     'allusers',
     'comment',
     'posts',
@@ -154,5 +155,13 @@ CORS_ALLOW_METHODS = (
     'PUT',
 )
 
+# Use nose to run all tests
+TEST_RUNNER = 'django_nose.NoseTestSuiteRunner'
+
+# Tell nose to measure coverage on the 'foo' and 'bar' apps
+NOSE_ARGS = [
+    '--with-coverage',
+    '--cover-package=posts.models,allusers.models,comment.models',
+]
 
 STATIC_URL = '/static/'
